pyproject.toml
[project]
name = "toto-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.100.0",
    "uvicorn>=0.23.0",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.13.1",
    "structlog>=25.4.0",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
]
